api/requirements.txt
google-genai==1.52.0
python-dotenv==1.2.1
supabase==2.24.0
fastapi==0.122.0
uvicorn==0.38.0
python-multipart==0.0.20
langchain==1.1.0
langchain-community==0.4.1
pypdf==6.4.0
# tensorflow==2.20.0
# tensorflow_hub==0.16.1
sentence-transformers==5.1.2
